Add csharp_nunit_test
I'm using the NUnitLite runner because it produces an executable which
works best with Bazel. I checked in some BUILD files for NUnitFramework
and NUnitLite, but those can be dropped once the default BUILD file for
NuGet packages gets added.

NOTE: we still aren't setting up the runfiles correctly, so "bazel test"
doesn't work correctly, but if you manually copy all the stuff out to a
folder it does.