Allow CSharpProviders to have neither out/refout
This is the easy part for #41.

The next step is to add csharp_library_set, which only has the deps
attribute. That's seems to be slightly non-trivial because our existing
helpers don't quite cut it.

The final piece of the puzzle is creating the @net//:StandardLibrary
target with a csharp_library_set that deps on @net48//:StandardLibrary
etc. That's easy but a bit tedious.