Wrap the dotnet executable to set default environment variables (#102)
To run `dotnet` cross-platform we need to ensure some environment are set. The problem though is that we must:

- Use an absolute path
- Have different paths for each OS (windows vs unix)

This commit implements a wrapper for dotnet (dotnetw) that ensures the environment variables are configured based on the operating system. Additionally it lays the groundwork for other rules to call dotnet build during execution. This is necessary for the work to support ResX files.

See also: #10, #98