Upgrade the Dotnet SDK to 3.1.100 from 3.0.100 (#103)
Upgrade to the latest version of the DotNET SDK.

Refactor the SDK URLs and version inputs to another file so updating in the future doesn't require modification in random files. 

Add note on how to update in the future, for others who are trying to keep our versions up to date.