Aggregate the bazel rules files for use in tooling (#112)
Aggregate all of the bazel sources together as using the model of `bzl_srcs`. This allows us to configure docs or tools on the bazel rule themselves. The top level is not accessible to any packages for now.

This PR aggregates all of the bazel sources to the top level using filegroups and the `bzl_srcs` model. This allows us to run tooling against our bazel rules files.